Inclusão de APIControlerBase, inicialização de banco de dados com alguns dados fake, ajuste para utilização do banco de dados em memória e ajuste na classe context, e inclusão da configuração para notificações de mensagen e erros que serão exibidos.

+++ Qualyteam.Application/Interfaces/IIndicadorMensalService.cs
@@ -8,10 +8,10 @@ namespace Qualyteam.Application.Interfaces
     public interface IIndicadorMensalService
     {
         Task<IEnumerable<IndicadorMensalViewModel>> Get();
-        Task<IndicadorMensalViewModel> GetById(long id);
+        Task<IndicadorMensalViewModel> GetById(int id);
         Task<IEnumerable<IndicadorMensalViewModel>> Search(FilterIndicadorMensalViewModel viewModel);
         Task<IndicadorMensalViewModel> Create(IndicadorMensalViewModel viewModel);
         Task<IndicadorMensalViewModel> Update(IndicadorMensalViewModel viewModel);
-        Task<int> Remove(long id);
+        Task<int> Remove(int id);
     }
 }
